BugFix
* Container에 최대용량(fullAmount) 이상으로 재료가 들어가는 문제 수정
* 재료를 소진했을 때 선택이 해제되지 않는 문제 수정
* 재료의 남은 용량보다 선택 용량(selectedPourAmount)가 많을 때 따라지는 문제 수정
* 일부 용량 변수의 type가 float로 수정되지 않은 문제 수정

+++ Assets/Scripts/ScriptableObejcts/Container.cs
@@ -2,7 +2,8 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public enum ContainerType {
+public enum ContainerType
+{
     Glass,
     Shaker,
     MixingGlass,
@@ -15,5 +16,5 @@ public class Container : ScriptableObject
     public ContainerType containerType;
     public Sprite containerSprite;
     public string containerName;
-    public int containerVolume;
+    public float containerVolume;
 }
